requirements.txt
flask
python-dotenv
google-genai
SpeechRecognition
pydub
reportlab
PyPDF2
gunicorn
